Implemented function to convert shape to raster based on specific extent
PLEASE NOTE THAT WE NEED TO SPECIFY PROPER ATTRIBUTE AND PARAMETERS WERE HARD CODED BASED ON PROJECTION 4326. If such projection is not acceptable, it is necessary to inform in_xsize, in_ysize, in_projection. Please note that there are several roudings, so minor differences should be expected. Since our goal is to compose rasters that will be used in a report, these minor mistakes are acceptable.